Added the correct version of sect20ex4
Uploaded the correct version of this file. The previous version did the file copy when the London office was open, not closed. This one uses an OR to make sure that we are excluding the operating hours. I also added an output to show that the file was not copied if the office is open because otherwise there is no indication of anything happening (or not happening.)