Added drills from Database/SQL course
These queries are for the drills from the Database SQL course (mostly the latter part of the course.) The primary folder is the 'LibrarySQL' folder.  The 'Library_drill.sql' file inside that folder creates and populates the library database. The other queries in that folder are the seven queries for that database drill (item 27).